+++ test/src/stringable_macro/empty_class_test.dart
@@ -1,10 +1,10 @@
-import 'package:data_class_macro/data_class_macro.dart';
+import 'package:data_class/data_class.dart';
 import 'package:test/test.dart';
 
 @Stringable()
 class EmptyClass {}
 
-void main() {  
+void main() {
   group(EmptyClass, () {
     test('toString is correct', () {
       expect(EmptyClass().toString(), equals('EmptyClass()'));